% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Passivevoice.R
\name{passivevoice}
\alias{passivevoice}
\title{Passive voice analyses}
\usage{
passivevoice(effect = NULL, sentence = NULL)
}
\arguments{
\item{effect}{the keyword to use as the "effect" (e.g. increase, decrease, promote)}

\item{sentence}{the sentence to be analyzed in active voice, parsed by coreNLP}
}
\value{
The drivers and the effect
}
\description{
Extracts the causal relation in passive sentences
For internal and experimental use
}
